+++ app/(admin)/admin/sources/page.tsx
@@ -7,18 +7,13 @@ import { Metadata } from 'next'
 import Link from 'next/link'
 import React from 'react'
 import Image from 'next/image'
-
-export const ASSESSMENT_STATUS_BEING_EVALUATED = 'being_evaluated'
-export const ASSESSMENT_STATUS_APPROVAL_NEEDED = 'approval_needed'
-export const ASSESSMENT_STATUS_PROOFREADING_NEEDED = 'proofreading_needed'
-export const ASSESSMENT_STATUS_APPROVED = 'approved'
-
-export const ASSESSMENT_STATUS_LABELS = {
-  [ASSESSMENT_STATUS_BEING_EVALUATED]: 've zpracování',
-  [ASSESSMENT_STATUS_APPROVAL_NEEDED]: 'ke kontrole',
-  [ASSESSMENT_STATUS_PROOFREADING_NEEDED]: 'ke korektuře',
-  [ASSESSMENT_STATUS_APPROVED]: 'schválený',
-}
+import {
+  ASSESSMENT_STATUS_APPROVAL_NEEDED,
+  ASSESSMENT_STATUS_APPROVED,
+  ASSESSMENT_STATUS_BEING_EVALUATED,
+  ASSESSMENT_STATUS_LABELS,
+  ASSESSMENT_STATUS_PROOFREADING_NEEDED,
+} from '@/libs/constants/assessment'
 
 export const metadata: Metadata = {
   title: getMetadataTitle('Seznam diskuzí', 'Administrace'),
